Refactor Container widget into modular structure
- Extract RippleState into container/ripple.rs module
- Extract AnimationState and helpers into container/animations.rs
- Extract scrollable functionality into container/scrollable.rs
- Add ScrollbarAxis enum to scroll.rs for unified scrollbar handling
- Add axis-parameterized methods to ScrollState for scrollbar calculations
- Deduplicate state layer resolution with resolve_state_value helper
- Deduplicate animation getters with get_animated_value helper
- Consolidate scrollbar creation with create_scrollbar_components

This reduces container.rs from ~2800 lines to ~1400 lines in mod.rs
with better separation of concerns.
